refactor(portal): drain Actions GraphQL docs to common (full-path mirror)
Move all 15 Actions operation docs to common/Teams/.../Actions keeping the full
subpath. The three GetSingleAction docs stay at distinct mirrored paths (no rename
needed under full-path). Repoint in-scene consumers + the standalone Kiosk scene.
No op-name renames; byte-identical relocation.
