+++ web/scenes/common/Teams/TeamId/Apps/AppId/Actions/ActionId/Kiosk/graphql/client/toggle-kiosk.generated.ts
