+++ web/scenes/common/Teams/TeamId/Apps/AppId/Actions/ActionId/Settings/UpdateAction/graphql/server/get-action-update-permissions.graphql
