+++ web/scenes/common/Teams/TeamId/Apps/AppId/Actions/ActionId/Settings/UpdateAction/graphql/server/update-action.graphql
