+++ web/scenes/common/Teams/TeamId/Apps/AppId/Actions/ActionId/Settings/page/graphql/client/get-single-action.graphql
