+++ web/scenes/common/Teams/TeamId/Apps/AppId/Actions/ActionId/page/ActionStatsGraph/graphql/client/fetch-action-stats.gql
