+++ web/scenes/common/Teams/TeamId/Apps/AppId/Actions/page/CreateActionModal/graphql/server/get-action-insert-permissions.graphql
